data model changes and observing more classes
So if either of you actually read the descriptions of my commits, the
changes that I made this evening are thus:
- The cars, people and loan objects themselves (not to be confused with
the list of the objects) now all inherit from Observed so their
properties can be monitored via their setter methods and calling
this.notifyObservers.
- I have changed the Library link so it now loads from the project
rather than the .jar. When building the project it will also build the
data jar.
- The observers are now split so that any changes to either the lists or
the objects themseves will call the category specific observer (Loans,
Cars or People/Staff).